Smoothing the streaming process
Frontend:
Clicking the "Next" button on the UI now causes the button to be disabled
for 1 second to prevent misclicks accidentally removing multiple groups
from the queue.

Queue length and the approximate amount of time remaining until the
queue is emptied is now above the queue table.

Backend:
Chatbot will now include the text of the !ign command when the Next
button is pressed in the UI.