Fix undefined parameter usage in `next`
The `next` method on QueueControls was still using the signature for
the version that was consuming an event, instead of being called by
QueueControls directly. The change to call the method directly didn't
take into account the lack of a supplied parameter, so was trying to
access an undefined variable. This change replaces the undefined `size`
with the `pop_size` data member.