Fix chat task being blocked by sleep
A command containing the BotAction::Wait action would wait until after
the duration of the sleep before sending any messages. This ended up
being due to tokio's task-granularity. Sleeping on the same task the IRC
client runs on caused the IRC client to be unable to flush until the wait
finished.