Beginning of axum migration
Axum is a less complicated web framework that, while less flexible than
Warp, will allow us to more quickly iterate thanks to the Tower
middleware. This isn't yet a 1-1 migration, because it doesn't support
the authentication flow. Once that is added, however, the main consumer
of it (the chatbot) is going to be removed in favor of supporting
external bots instead.