Remove chatbot implementation
The chatbot was beginning to become a huge drag on implementation
momentum. Rather than keep trying to prop up an interface that was
becoming exceedingly complicated, I'm opting instead to remove it so I
can focus on the core experience of the application. Chat integration
can instead rely on StreamElements bot implementation.